// src/types/types.ts
export interface AppConfig {
  appId?: string
  appName: string
  appNameShort: string
  appStore: string
  backupAccountSite: string
  configName: string
  darkTheme: Theme
  defaultWallets: string[]
  knowledgeBase: string
  lightTheme: Theme
  notificationServers: string[]
  phoneNumber: string
  referralServers?: string[]
  supportsEdgeLogin: boolean
  supportEmail: string
  supportSite: string
  termsOfServiceSite: string
  website: string
  disableSwaps?: boolean
  extraTab?: {
    webviewUrl: string
    tabType: 'edgeProvider' | 'webview'
    tabTitleKey: LocaleStringKey
    extraTabBarIconFont: string
    extraTabBarIconName: string
  }
  extraTab2?: {
    tabTitleKey: LocaleStringKey,
    extraTab2BarIconFont: string,
    extraTab2BarIconName: string
  }
}